requirements.txt
streamlit
langchain
langchain-openai
langchain-google-genai
langchain-anthropic
langchain-community
notion-client
python-dotenv
pydantic
httpx
tenacity
